Outline for conversions
Right now there is a lot of code in whatever_ms or whatever_us

and I would much rather have it be in ns always, and be able
to specify what the units were via whatever = 3.2ms.

I found the old units command, but really, I just wanted an
off the shelf lookup table for mbits/kibibits and ms to ns
and was too lazy to write it yet.