Yay: I now know the sizes of key structures
In particular, I was wrong about the interface name size,
it is actually 16 characters including the NULL, and hidden
in a kernel define for if.h as the so 80s IFNAMSIZ define.

FIXME: Go fix this in the rust dummy0 thing.

Interface names are longer on windows, and I have seen desparate
individuals attempt to use BASE64 for it in linux. I am not really
sure how to cope with containers. To me a container is just eth0.